[CDEC-457] Reunite orphans in Pos.Core.Block.Union.Instances
