+++ core/src/Pos/Core/Block/Genesis/Instances.hs
@@ -8,46 +8,25 @@ module Pos.Core.Block.Genesis.Instances
 
 import           Universum
 
-import           Formatting (bprint, build, int, sformat, stext, (%))
+import           Formatting (bprint, build, sformat, stext, (%))
 import qualified Formatting.Buildable as Buildable
 import           Serokell.Util (Color (Magenta), colorize)
 
-import           Pos.Core.Block.Blockchain (GenericBlock (..),
-                     GenericBlockHeader (..), gbHeader, gbhConsensus)
-import           Pos.Core.Block.Genesis.Lens (gcdDifficulty, gcdEpoch)
-import           Pos.Core.Block.Genesis.Types (GenesisBody (..),
-                     GenesisConsensusData (..))
+import           Pos.Core.Block.Blockchain (GenericBlock (..), gbHeader,
+                     gbhConsensus)
+import           Pos.Core.Block.Genesis.Lens (gcdEpoch)
+import           Pos.Core.Block.Genesis.Types (GenesisBody (..))
 import           Pos.Core.Block.Union.Types (BlockHeader (..), GenesisBlock,
-                     GenesisBlockHeader, HasHeaderHash (..), HeaderHash,
-                     IsGenesisHeader, IsHeader, blockHeaderHash)
-import           Pos.Core.Common (HasDifficulty (..), slotLeadersF)
-import           Pos.Core.Slotting (EpochOrSlot (..), HasEpochIndex (..),
-                     HasEpochOrSlot (..))
-import           Pos.Crypto (hashHexF)
+                     GenesisBlockHeader, HasHeaderHash (..), IsGenesisHeader,
+                     IsHeader, blockHeaderHash)
+import           Pos.Core.Common (slotLeadersF)
+import           Pos.Core.Slotting (HasEpochIndex (..), HasEpochOrSlot (..))
 
 instance NFData GenesisBlock
 
 ----------------------------------------------------------------------------
 -- Buildable
 ----------------------------------------------------------------------------
-
-instance Buildable GenesisBlockHeader where
-    build gbh@UnsafeGenericBlockHeader {..} =
-        bprint
-            ("GenesisBlockHeader:\n"%
-             "    hash: "%hashHexF%"\n"%
-             "    previous block: "%hashHexF%"\n"%
-             "    epoch: "%build%"\n"%
-             "    difficulty: "%int%"\n"
-            )
-            gbhHeaderHash
-            _gbhPrevBlock
-            _gcdEpoch
-            _gcdDifficulty
-      where
-        gbhHeaderHash :: HeaderHash
-        gbhHeaderHash = blockHeaderHash $ BlockHeaderGenesis gbh
-        GenesisConsensusData {..} = _gbhConsensus
 
 instance Buildable GenesisBlock where
     build UnsafeGenericBlock {..} =
@@ -68,12 +47,6 @@ instance Buildable GenesisBlock where
 instance HasEpochIndex GenesisBlock where
     epochIndexL = gbHeader . gbhConsensus . gcdEpoch
 
-instance HasEpochIndex GenesisBlockHeader where
-    epochIndexL = gbhConsensus . gcdEpoch
-
-instance HasEpochOrSlot GenesisBlockHeader where
-    getEpochOrSlot = EpochOrSlot . Left . _gcdEpoch . _gbhConsensus
-
 instance HasEpochOrSlot GenesisBlock where
     getEpochOrSlot = getEpochOrSlot . _gbHeader
 
@@ -82,15 +55,6 @@ instance HasHeaderHash GenesisBlockHeader where
 
 instance HasHeaderHash GenesisBlock where
     headerHash = blockHeaderHash . BlockHeaderGenesis . _gbHeader
-
-instance HasDifficulty GenesisConsensusData where
-    difficultyL = gcdDifficulty
-
-instance HasDifficulty GenesisBlockHeader where
-    difficultyL = gbhConsensus . difficultyL
-
-instance HasDifficulty GenesisBlock where
-    difficultyL = gbHeader . difficultyL
 
 instance IsHeader GenesisBlockHeader
 instance IsGenesisHeader GenesisBlockHeader
