+++ core/src/Pos/Core/Block/Genesis/Lens.hs
@@ -44,9 +44,9 @@ import           Pos.Core.Block.Blockchain (gbBody, gbExtra, gbHeader,
                      gbPrevBlock, gbhBodyProof, gbhConsensus, gbhExtra,
                      gbhPrevBlock)
 import           Pos.Core.Block.Genesis.Types (GenesisBody (..),
-                     GenesisBodyAttributes, GenesisConsensusData (..),
-                     GenesisExtraBodyData (..), GenesisExtraHeaderData (..),
-                     GenesisHeaderAttributes, GenesisProof (..))
+                     GenesisBodyAttributes, GenesisExtraBodyData (..),
+                     GenesisExtraHeaderData (..), GenesisHeaderAttributes,
+                     GenesisProof (..), gcdDifficulty, gcdEpoch)
 import           Pos.Core.Block.Union.Types (GenesisBlock, GenesisBlockHeader,
                      HeaderHash)
 import           Pos.Core.Common (ChainDifficulty, SlotLeaders)
@@ -62,8 +62,6 @@ makeLenses ''GenesisExtraBodyData
 ----------------------------------------------------------------------------
 -- ConsensusData
 ----------------------------------------------------------------------------
-
-makeLenses 'GenesisConsensusData
 
 ----------------------------------------------------------------------------
 -- GenesisBlockHeader
