+++ core/src/Pos/Core/Block/Main.hs
@@ -3,10 +3,8 @@
 
 module Pos.Core.Block.Main
       ( module Pos.Core.Block.Main.Instances
-      , module Pos.Core.Block.Main.Lens
       , module Pos.Core.Block.Main.Types
       ) where
 
 import           Pos.Core.Block.Main.Instances
-import           Pos.Core.Block.Main.Lens
 import           Pos.Core.Block.Main.Types
