+++ core/src/Pos/Core/Block/Main/Instances.hs
@@ -12,51 +12,22 @@ import           Formatting (bprint, build, int, stext, (%))
 import qualified Formatting.Buildable as Buildable
 import           Serokell.Util (Color (Magenta), colorize, listJson)
 
-import           Pos.Core.Block.Blockchain (GenericBlock (..),
-                     GenericBlockHeader (..))
-import           Pos.Core.Block.Main.Lens (mainBlockBlockVersion,
-                     mainBlockDifficulty, mainBlockSlot,
-                     mainBlockSoftwareVersion, mainHeaderBlockVersion,
-                     mainHeaderDifficulty, mainHeaderLeaderKey, mainHeaderSlot,
-                     mainHeaderSoftwareVersion, mbTxs, mcdDifficulty,
-                     mehBlockVersion, mehSoftwareVersion)
+import           Pos.Core.Block.Blockchain (GenericBlock (..))
 import           Pos.Core.Block.Main.Types (MainBody (..),
                      MainExtraHeaderData (..))
 import           Pos.Core.Block.Union.Types (BlockHeader (..),
-                     HasHeaderHash (..), HeaderHash, IsHeader,
-                     IsMainHeader (..), MainBlock, MainBlockHeader,
-                     MainConsensusData (..), blockHeaderHash)
+                     HasHeaderHash (..), IsHeader, IsMainHeader (..),
+                     MainBlock, MainBlockHeader, MainConsensusData (..),
+                     blockHeaderHash, mainBlockBlockVersion, mainBlockSlot,
+                     mainBlockSoftwareVersion, mainHeaderBlockVersion,
+                     mainHeaderLeaderKey, mainHeaderSlot,
+                     mainHeaderSoftwareVersion, mbTxs, mcdDifficulty,
+                     mehBlockVersion, mehSoftwareVersion)
 import           Pos.Core.Common (HasDifficulty (..))
-import           Pos.Core.Slotting (EpochOrSlot (..), HasEpochIndex (..),
-                     HasEpochOrSlot (..), slotIdF)
+import           Pos.Core.Slotting (HasEpochIndex (..), HasEpochOrSlot (..))
 import           Pos.Core.Update (HasBlockVersion (..), HasSoftwareVersion (..))
-import           Pos.Crypto (hashHexF)
 
 instance NFData MainBlock
-
-instance Buildable MainBlockHeader where
-    build gbh@UnsafeGenericBlockHeader {..} =
-        bprint
-            ("MainBlockHeader:\n"%
-             "    hash: "%hashHexF%"\n"%
-             "    previous block: "%hashHexF%"\n"%
-             "    slot: "%slotIdF%"\n"%
-             "    difficulty: "%int%"\n"%
-             "    leader: "%build%"\n"%
-             "    signature: "%build%"\n"%
-             build
-            )
-            gbhHeaderHash
-            _gbhPrevBlock
-            _mcdSlot
-            _mcdDifficulty
-            _mcdLeaderKey
-            _mcdSignature
-            _gbhExtra
-      where
-        gbhHeaderHash :: HeaderHash
-        gbhHeaderHash = blockHeaderHash $ BlockHeaderMain gbh
-        MainConsensusData {..} = _gbhConsensus
 
 instance Buildable MainBlock where
     build UnsafeGenericBlock {..} =
@@ -84,12 +55,6 @@ instance Buildable MainBlock where
 instance HasEpochIndex MainBlock where
     epochIndexL = mainBlockSlot . epochIndexL
 
-instance HasEpochIndex MainBlockHeader where
-    epochIndexL = mainHeaderSlot . epochIndexL
-
-instance HasEpochOrSlot MainBlockHeader where
-    getEpochOrSlot = EpochOrSlot . Right . view mainHeaderSlot
-
 instance HasEpochOrSlot MainBlock where
     getEpochOrSlot = getEpochOrSlot . _gbHeader
 
@@ -101,12 +66,6 @@ instance HasHeaderHash MainBlock where
 
 instance HasDifficulty MainConsensusData where
     difficultyL = mcdDifficulty
-
-instance HasDifficulty MainBlockHeader where
-    difficultyL = mainHeaderDifficulty
-
-instance HasDifficulty MainBlock where
-    difficultyL = mainBlockDifficulty
 
 instance HasBlockVersion MainExtraHeaderData where
     blockVersionL = mehBlockVersion
