+++ core/src/Pos/Core/Block/Union.hs
@@ -1,7 +1,7 @@
 module Pos.Core.Block.Union
-       ( module Pos.Core.Block.Union.Instances
+       ( module Pos.Core.Block.Union.ComponentBlock
        , module Pos.Core.Block.Union.Types
        ) where
 
-import           Pos.Core.Block.Union.Instances
+import           Pos.Core.Block.Union.ComponentBlock
 import           Pos.Core.Block.Union.Types
